requirements.txt
openai
numpy
pandas
matplotlib
seaborn
datasets
faiss-cpu
llama-index-core
llama-index-embeddings-ollama
llama-index-vector-stores-faiss
